% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_functions.R
\docType{package}
\name{appengine_googleAuthR}
\alias{appengine_googleAuthR}
\alias{appengine_googleAuthR-package}
\title{Google App Engine Admin API
Provisions and manages App Engine applications.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:27:48
filename: /Users/mark/dev/R/autoGoogleAPI/googleappenginev1beta.auto/R/appengine_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/appengine.admin
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/cloud-platform
}
}
