% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{Application}
\alias{Application}
\title{Application Object}
\usage{
Application(codeBucket = NULL, defaultBucket = NULL, dispatchRules = NULL,
  gcrDomain = NULL, name = NULL, defaultCookieExpiration = NULL,
  id = NULL, locationId = NULL, servingStatus = NULL,
  defaultHostname = NULL, authDomain = NULL, iap = NULL)
}
\arguments{
\item{codeBucket}{Google Cloud Storage bucket that can be used for storing files associated with this application}

\item{defaultBucket}{Google Cloud Storage bucket that can be used by this application to store content}

\item{dispatchRules}{HTTP path dispatch rules for requests to the application that do not explicitly target a service or version}

\item{gcrDomain}{The Google Container Registry domain used for storing managed build docker images for this application}

\item{name}{Full path to the Application resource in the API}

\item{defaultCookieExpiration}{Cookie expiration policy for this application}

\item{id}{Identifier of the Application resource}

\item{locationId}{Location from which this application will be run}

\item{servingStatus}{Serving status of this application}

\item{defaultHostname}{Hostname used to reach this application, as resolved by App Engine}

\item{authDomain}{Google Apps authentication domain that controls which users can access this application}

\item{iap}{No description}
}
\value{
Application object
}
\description{
Application Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Application resource contains the top-level configuration of an App Engine application.
}
\seealso{
Other Application functions: \code{\link{apps.create}},
  \code{\link{apps.patch}}
}
