% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{DebugInstanceRequest}
\alias{DebugInstanceRequest}
\title{DebugInstanceRequest Object}
\usage{
DebugInstanceRequest(sshKey = NULL)
}
\arguments{
\item{sshKey}{Public SSH key to add to the instance}
}
\value{
DebugInstanceRequest object
}
\description{
DebugInstanceRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for Instances.DebugInstance.
}
