% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{FileInfo}
\alias{FileInfo}
\title{FileInfo Object}
\usage{
FileInfo(sha1Sum = NULL, mimeType = NULL, sourceUrl = NULL)
}
\arguments{
\item{sha1Sum}{The SHA1 hash of the file, in hex}

\item{mimeType}{The MIME type of the file}

\item{sourceUrl}{URL source to use to fetch this file}
}
\value{
FileInfo object
}
\description{
FileInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Single source file that is part of the version to be deployed. Each source file that is deployed must be specified separately.
}
