% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{HealthCheck}
\alias{HealthCheck}
\title{HealthCheck Object}
\usage{
HealthCheck(host = NULL, restartThreshold = NULL, healthyThreshold = NULL,
  checkInterval = NULL, timeout = NULL, unhealthyThreshold = NULL,
  disableHealthCheck = NULL)
}
\arguments{
\item{host}{Host header to send when performing an HTTP health check}

\item{restartThreshold}{Number of consecutive failed health checks required before an instance is restarted}

\item{healthyThreshold}{Number of consecutive successful health checks required before receiving traffic}

\item{checkInterval}{Interval between health checks}

\item{timeout}{Time before the health check is considered failed}

\item{unhealthyThreshold}{Number of consecutive failed health checks required before removing traffic}

\item{disableHealthCheck}{Whether to explicitly disable health checks for this instance}
}
\value{
HealthCheck object
}
\description{
HealthCheck Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Health checking configuration for VM instances. Unhealthy instances are killed and replaced with new instances. Only applicable for instances in App Engine flexible environment.
}
