% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{IdentityAwareProxy}
\alias{IdentityAwareProxy}
\title{IdentityAwareProxy Object}
\usage{
IdentityAwareProxy(oauth2ClientSecret = NULL, oauth2ClientId = NULL,
  oauth2ClientSecretSha256 = NULL, enabled = NULL)
}
\arguments{
\item{oauth2ClientSecret}{OAuth2 client secret to use for the authentication flow}

\item{oauth2ClientId}{OAuth2 client ID to use for the authentication flow}

\item{oauth2ClientSecretSha256}{Hex-encoded SHA-256 hash of the client secret}

\item{enabled}{Whether the serving infrastructure will authenticate and authorize all incoming requests}
}
\value{
IdentityAwareProxy object
}
\description{
IdentityAwareProxy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Identity-Aware Proxy
}
