% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{NetworkUtilization}
\alias{NetworkUtilization}
\title{Google App Engine Admin API Objects 
Provisions and manages App Engine applications.}
\usage{
NetworkUtilization(targetSentBytesPerSecond = NULL,
  targetSentPacketsPerSecond = NULL, targetReceivedBytesPerSecond = NULL,
  targetReceivedPacketsPerSecond = NULL)
}
\arguments{
\item{targetSentBytesPerSecond}{Target bytes sent per second}

\item{targetSentPacketsPerSecond}{Target packets sent per second}

\item{targetReceivedBytesPerSecond}{Target bytes received per second}

\item{targetReceivedPacketsPerSecond}{Target packets received per second}
}
\value{
NetworkUtilization object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:27:48
filename: /Users/mark/dev/R/autoGoogleAPI/googleappenginev1beta.auto/R/appengine_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
NetworkUtilization Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Target scaling by network usage. Only applicable for VM runtimes.
}
