% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{OperationMetadata}
\alias{OperationMetadata}
\title{OperationMetadata Object}
\usage{
OperationMetadata(insertTime = NULL, target = NULL, user = NULL,
  method = NULL, endTime = NULL, operationType = NULL)
}
\arguments{
\item{insertTime}{Timestamp that this operation was created}

\item{target}{Name of the resource that this operation is acting on}

\item{user}{User who requested this operation}

\item{method}{API method that initiated this operation}

\item{endTime}{Timestamp that this operation completed}

\item{operationType}{Type of this operation}
}
\value{
OperationMetadata object
}
\description{
OperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata for the given google.longrunning.Operation.
}
