% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{OperationMetadataExperimental}
\alias{OperationMetadataExperimental}
\title{OperationMetadataExperimental Object}
\usage{
OperationMetadataExperimental(endTime = NULL, user = NULL, target = NULL,
  method = NULL, insertTime = NULL)
}
\arguments{
\item{endTime}{Time that this operation completed}

\item{user}{User who requested this operation}

\item{target}{Name of the resource that this operation is acting on}

\item{method}{API method that initiated this operation}

\item{insertTime}{Time that this operation was created}
}
\value{
OperationMetadataExperimental object
}
\description{
OperationMetadataExperimental Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata for the given google.longrunning.Operation.
}
