% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{OperationMetadataV1Beta5}
\alias{OperationMetadataV1Beta5}
\title{OperationMetadataV1Beta5 Object}
\usage{
OperationMetadataV1Beta5(user = NULL, target = NULL, method = NULL,
  insertTime = NULL, endTime = NULL)
}
\arguments{
\item{user}{User who requested this operation}

\item{target}{Name of the resource that this operation is acting on}

\item{method}{API method name that initiated this operation}

\item{insertTime}{Timestamp that this operation was created}

\item{endTime}{Timestamp that this operation completed}
}
\value{
OperationMetadataV1Beta5 object
}
\description{
OperationMetadataV1Beta5 Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata for the given google.longrunning.Operation.
}
