% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{RepairApplicationRequest}
\alias{RepairApplicationRequest}
\title{RepairApplicationRequest Object}
\usage{
RepairApplicationRequest()
}
\value{
RepairApplicationRequest object
}
\description{
RepairApplicationRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for 'Applications.RepairApplication'.
}
\seealso{
Other RepairApplicationRequest functions: \code{\link{apps.repair}}
}
