% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{Service}
\alias{Service}
\title{Service Object}
\usage{
Service(id = NULL, name = NULL, split = NULL)
}
\arguments{
\item{id}{Relative name of the service within the application}

\item{name}{Full path to the Service resource in the API}

\item{split}{Mapping that defines fractional HTTP traffic diversion to different versions within the service}
}
\value{
Service object
}
\description{
Service Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Service resource is a logical component of an application that can share state and communicate in a secure fashion with other services. For example, an application that handles customer requests might include separate services to handle tasks such as backend data analysis or API requests from mobile devices. Each service has a collection of versions that define a specific set of code used to implement the functionality of that service.
}
