% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{UrlMap}
\alias{UrlMap}
\title{UrlMap Object}
\usage{
UrlMap(urlRegex = NULL, login = NULL, apiEndpoint = NULL,
  staticFiles = NULL, redirectHttpResponseCode = NULL,
  securityLevel = NULL, authFailAction = NULL, script = NULL)
}
\arguments{
\item{urlRegex}{URL prefix}

\item{login}{Level of login required to access this resource}

\item{apiEndpoint}{Uses API Endpoints to handle requests}

\item{staticFiles}{Returns the contents of a file, such as an image, as the response}

\item{redirectHttpResponseCode}{30x code to use when performing redirects for the secure field}

\item{securityLevel}{Security (HTTPS) enforcement for this URL}

\item{authFailAction}{Action to take when users access resources that require authentication}

\item{script}{Executes a script to handle the request that matches this URL pattern}
}
\value{
UrlMap object
}
\description{
UrlMap Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
URL pattern and description of how the URL should be handled. App Engine can handle URLs by executing application code or by serving static files uploaded with the version, such as images, CSS, or JavaScript.
}
