% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{ZipInfo}
\alias{ZipInfo}
\title{ZipInfo Object}
\usage{
ZipInfo(sourceUrl = NULL, filesCount = NULL)
}
\arguments{
\item{sourceUrl}{URL of the zip file to deploy from}

\item{filesCount}{An estimate of the number of files in a zip for a zip deployment}
}
\value{
ZipInfo object
}
\description{
ZipInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The zip file information for a zip deployment.
}
